// src/api/gen/_model/listUserRankingsSort.gen.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * TMT API
 * 또맛또(TMT) API
 * OpenAPI spec version: v1
 */

export type ListUserRankingsSort = (typeof ListUserRankingsSort)[keyof typeof ListUserRankingsSort];

export const ListUserRankingsSort = {
  reviewCount: "reviewCount",
  sharedReviewCount: "sharedReviewCount",
} as const;